+++ uniCloud-aliyun/cloudfunctions/router/service/user/kh/unbindQQ.js
@@ -1,8 +1,7 @@
 module.exports = {
   /**
-   * 解绑支付宝
-   * @url user/kh/unbindAlipay 前端调用的url参数地址
-   * @description 当前登录用户解绑支付宝
+   * 解绑QQ
+   * @url user/kh/unbindQQ 前端调用的url参数地址
 	 * res 返回参数说明
 	 * @param {Number} code 错误码，0表示成功
 	 * @param {String} msg 详细信息
@@ -14,10 +13,10 @@ module.exports = {
 		let { uid } = data;
 		let res = { code : 0, msg : '' };
 		if(vk.pubfn.isNullAll(userInfo.mobile, userInfo.username)){
-			return { code : -1, msg : '为了您的账号安全，请绑定手机号后再解绑支付宝！' };
+			return { code : -1, msg : '为了您的账号安全，请绑定手机号后再解绑QQ！' };
 		}
 		// 解绑
-		res = await uniID.unbindAlipay(uid);
+		res = await uniID.unbindQQ(uid);
 		// 业务逻辑结束-----------------------------------------------------------
 		return res;
 	}