+++ uni_modules/vk-unicloud/vk_modules/vk-unicloud-page/libs/function/index.js
@@ -663,6 +663,22 @@ pubfn.checkDataExpText = function (data={},expText) {
 	return checkKey;
 };
 
+/**
+ * 判断变量是否是数组
+ * vk.pubfn.isArray(value);
+ */
+pubfn.isArray = function (value) {
+	return Object.prototype.toString.call(value) === "[object Array]" ? true : false;
+};
+/**
+ * 判断变量是否是对象
+ * vk.pubfn.isObject(value);
+ */
+pubfn.isObject = function (value) {
+	return Object.prototype.toString.call(value) === "[object Object]" ? true : false;
+};
+
+
 
 /**
  * 计算运费